Added Certificate Alias Max Size (128 Characters). Moved all validators in a new package (org.eclipse.kura.web.client.ui.validator)
