+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/security/KeyPairTabUi.java
@@ -15,11 +15,12 @@ package org.eclipse.kura.web.client.ui.security;
 import java.util.List;
 
 import org.eclipse.kura.web.client.messages.Messages;
-import org.eclipse.kura.web.client.ui.NotEmptyValidator;
-import org.eclipse.kura.web.client.ui.PEMValidator;
-import org.eclipse.kura.web.client.ui.PKCS8Validator;
-import org.eclipse.kura.web.client.ui.StringNotInListValidator;
 import org.eclipse.kura.web.client.ui.Tab;
+import org.eclipse.kura.web.client.ui.validator.NotEmptyValidator;
+import org.eclipse.kura.web.client.ui.validator.PEMValidator;
+import org.eclipse.kura.web.client.ui.validator.PKCS8Validator;
+import org.eclipse.kura.web.client.ui.validator.StringLengthValidator;
+import org.eclipse.kura.web.client.ui.validator.StringNotInListValidator;
 import org.eclipse.kura.web.client.util.request.RequestQueue;
 import org.eclipse.kura.web.shared.service.GwtCertificatesService;
 import org.eclipse.kura.web.shared.service.GwtCertificatesServiceAsync;
@@ -40,6 +41,8 @@ import com.google.gwt.user.client.ui.HTMLPanel;
 import com.google.gwt.user.client.ui.Widget;
 
 public class KeyPairTabUi extends Composite implements Tab {
+
+    private static final int ALIAS_MAX_LENGTH = 128;
 
     public enum Type {
         KEY_PAIR,
@@ -141,6 +144,9 @@ public class KeyPairTabUi extends Composite implements Tab {
         NotEmptyValidator notEmptyValidator = new NotEmptyValidator(MSGS.formRequiredParameter());
 
         this.storageAliasInput.addValidator(notEmptyValidator);
+        this.storageAliasInput.addValidator(new StringLengthValidator(ALIAS_MAX_LENGTH,
+                MSGS.certificateAliasMaxLength(String.valueOf(ALIAS_MAX_LENGTH))));
+        this.storageAliasInput.setMaxLength(ALIAS_MAX_LENGTH);
         this.storageAliasInput.addValidator(new StringNotInListValidator(usedAliases, MSGS.certificateAliasUsed()));
 
         this.certificateInput.addValidator(notEmptyValidator);
