+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/validator/PredicateValidator.java
@@ -10,7 +10,7 @@
  * Contributors:
  *  Eurotech
  *******************************************************************************/
-package org.eclipse.kura.web.client.ui;
+package org.eclipse.kura.web.client.ui.validator;
 
 import java.util.Collections;
 import java.util.List;
