+++ kura/org.eclipse.kura.web2/src/main/resources/org/eclipse/kura/web/client/messages/Messages.properties
@@ -275,6 +275,7 @@ settingsReloadStartupFingerprintDescription=Allows to change the startup command
 
 certificateAlias=Alias
 certificateAliasUsed=Certificate Alias already used.
+certificateAliasMaxLength=Alias must be at most {0} characters
 certificateKeystoreName=Keystore Service Name
 certificateKind=Type
 certificateValidityStart=Valid from
